requirements.txt
altair==5.5.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
click==8.2.1
gitdb==4.0.12
GitPython==3.1.44
google==3.0.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
learn==1.0.0
narwhals==1.43.0
numpy==2.3.0
protobuf==6.31.1
pyarrow==20.0.0
pydeck==0.9.1
referencing==0.36.2
rpds-py==0.25.1
scikit-base==0.12.3
scikit-learn==1.7.0
scipy==1.15.3
setuptools==80.9.0
sk==0.0.1
sktime==0.37.0
smmap==5.0.2
soupsieve==2.7
streamlit==1.45.1
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
typing_extensions==4.14.0
watchdog==6.0.0
wheel==0.45.1
